+++ datatig/__main__.py
@@ -1,0 +1,4 @@
+import datatig.cli
+
+if __name__ == "__main__":
+    datatig.cli.main()